requirements.txt
boto3
botocore
requests
pydantic>=2
strands-agents
strands-agents-tools
bedrock-agentcore
bedrock-agentcore-starter-toolkit
mcp
jsonschema
